# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /billing/apple/transaction:
    post:
      operationId: verifyAppleTransaction
      tags: [billing]
      summary: Verify and claim an App Store subscription transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleVerifyInput"
      responses:
        "201":
          description: Verified Apple subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppleVerification"
  /billing/apple/restore:
    post:
      operationId: restoreAppleTransactions
      tags: [billing]
      summary: Restore verified App Store subscription transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleRestoreInput"
      responses:
        "200":
          description: Restored Apple subscriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppleRestoreResponse"
  /apple/webhook:
    post:
      operationId: receiveAppleWebhook
      tags: [billing]
      summary: Receive an App Store Server Notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleWebhookInput"
      responses:
        "200":
          description: Notification received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppleWebhookResponse"
components:
  schemas:
    AppleWebhookInput:
      type: object
      required: [signedPayload]
      properties:
        signedPayload:
          type: string
          minLength: 1
    AppleWebhookResponse:
      type: object
      required: [received]
      properties:
        received:
          type: boolean
          const: true
    AppleVerifyInput:
      type: object
      required: [signedTransaction]
      properties:
        signedTransaction:
          type: string
          minLength: 1
    AppleRestoreInput:
      type: object
      required: [signedTransactions]
      properties:
        signedTransactions:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: string
            minLength: 1
    AppleVerification:
      type: object
      required: [plan, originalTransactionId]
      properties:
        plan:
          type: string
          enum: [weekly, monthly, yearly]
        originalTransactionId:
          type: string
    AppleRestoreResponse:
      type: object
      required: [restored]
      properties:
        restored:
          type: array
          items:
            $ref: "#/components/schemas/AppleVerification"
...