# lib/api-zod/src/generated/types/appleRestoreInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppleRestoreInput {
  /**
     * @minItems 1
     * @maxItems 100
     * @items.minLength 1
     */
  signedTransactions: string[];
}