# lib/api-zod/src/generated/types/appleRestoreResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppleVerification } from './appleVerification';

export interface AppleRestoreResponse {
  restored: AppleVerification[];
}