# lib/api-zod/src/generated/types/appleVerification.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AppleVerificationPlan } from './appleVerificationPlan';

export interface AppleVerification {
  plan: AppleVerificationPlan;
  originalTransactionId: string;
}