# lib/api-zod/src/generated/types/appleWebhookResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export const AppleWebhookResponseValue = {
  received: true,
} as const;
export type AppleWebhookResponse = typeof AppleWebhookResponseValue;